commit after datatype missmatch with data base resolved.

+++ Test/TracyBodine_P0_BankingApp/src/main/java/com/revature/p0/services/AccountService.java
@@ -33,7 +33,7 @@ public class AccountService {
 	private boolean isAccountValid(Account newAccount) {
 		
 		if(newAccount == null) return false;
-		if(newAccount.getAccountName() == null || newAccount.getAccountName().trim().equals("")|| Integer.valueOf(newAccount.getAccountBalance()) < 0) return false;
+		if(newAccount.getAccountName() == null || newAccount.getAccountName().trim().equals("")|| Double.valueOf(newAccount.getAccountBalance()) < 0) return false;
 		return newAccount.getAccountName() != null || !newAccount.getAccountName().trim().equals("");
 	}
